Front / Refactoring: project repository to spring boot repository
