+++ src/test/javascript/spec/springboot/secondary/ProjectRepository.spec.ts
@@ -47,34 +47,6 @@ describe('ProjectRepository', () => {
     expect(payload).toEqual<RestProject>(expectedRestProject);
   });
 
-  it('should add SpringBoot', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBoot(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add SpringBoot MVC with Tomcat', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootMvcTomcat(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/mvc/web/tomcat');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
   it('should add Frontend Maven Plugin', () => {
     const axiosHttpStub = stubAxiosHttp();
     axiosHttpStub.post.resolves();
@@ -86,380 +58,6 @@ describe('ProjectRepository', () => {
     const expectedRestProject: RestProject = toRestProject(project);
     const [uri, payload] = axiosHttpStub.post.getCall(0).args;
     expect(uri).toBe('api/frontend-maven-plugin');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add Ippon Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerIppon(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/ippon');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JHipster v2 Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerJHipsterV2(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/jhipster-v2');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JHipster v3 Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerJHipsterV3(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/jhipster-v3');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JHipster v7 Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerJHipsterV7(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/jhipster-v7');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JHipster v7 React Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerJHipsterV7React(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/jhipster-v7-react');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JHipster v7 Vue Banner', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootBannerJHipsterV7Vue(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/spring-boot/banner/jhipster-v7-vue');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add Postgres', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addPostgres(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/postgresql');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add Mysql', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addMySQL(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/mysql');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add MariaDB', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addMariaDB(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/mariadb');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add MongoDB', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addMongoDB(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/mongodb');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should migrate flyway init', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootFlywayInit(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/migration/flyway/init');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should migrate flyway user', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootFlywayUser(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/migration/flyway/user');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should migrate liquibase init', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootLiquibaseInit(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/migration/liquibase/init');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should migrate liquibase user', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootLiquibaseUser(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/databases/migration/liquibase/user');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add JWT', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addJWT(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/mvc/security/jwt');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add Basic Auth JWT', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addBasicAuthJWT(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/mvc/security/jwt/basic-auth');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add Oauth2', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addOauth2(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/mvc/security/oauth2');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add SpringdocJWT', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringdocJWT(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/mvc/springdoc/init-with-security-jwt');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add asynchronous execution', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootAsync(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/async');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add devtools dependencies', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootDevtoolsDependencies(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/devtools');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add dockerfile', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootDockerfile(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/docker/dockerfile');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add docker jib', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringBootDockerJib(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/docker/jib');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add client config on spring cloud', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringCloudConfigClient(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/spring-cloud/config-client');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add spring cloud consul config', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringCloudConsul(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/spring-cloud/consul');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add eureka on spring cloud', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSpringCloudEureka(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/spring-cloud/eureka-client');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-
-  it('should add Ehcache with Java configuration', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addEhcacheWithJavaConf(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/cache/ehcache/java-configuration');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add Ehcache with XML', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addEhcacheWithXML(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/cache/ehcache/xml-configuration');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-  });
-  it('should add simple cache', () => {
-    const axiosHttpStub = stubAxiosHttp();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub);
-    const project: Project = createProject({ folder: 'folder/path' });
-
-    projectRepository.addSimpleCache(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('/api/servers/spring-boot/cache/simple');
     expect(payload).toEqual<RestProject>(expectedRestProject);
   });
 });
